mejoramiento del main
